requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
web3==7.13.0
python-dotenv==1.1.1
sentence-transformers==5.1.0
torch==2.8.0
python-multipart==0.0.20
jinja2==3.1.6
pypdf==5.9.0
